Create sidebar-single-staff.php
This will allow staff,team,department,section, or other groups of users to be displayed as a sidebar on a staff members page.

Will only appear if a Staff member has staff assigned to them via the attributes box of the specified Team member's Staff page. 

For example a Children's Pastor can be assigned to the Senior Pastor's staff.